fix warn

+++ window_scene/interfaces/kits/napi/screen_session_manager/js_screen_utils.cpp
@@ -196,12 +196,12 @@ bool ConvertDMRectFromJs(napi_env env, napi_value jsObject, DMRect& rect)
         rect.posY_ = top;
     }
     if (GetType(env, jsWidth) != napi_undefined) {
-        uint32_t width;
+        int32_t width;
         if (!ConvertFromJsValue(env, jsWidth, width)) {
             WLOGFE("[NAPI]Failed to convert parameter to width");
             return false;
         }
-        rect.width_ = width;
+        rect.width_ = static_cast<uint32_t>(width);
     }
     if (GetType(env, jsHeight) != napi_undefined) {
         int32_t height;
@@ -209,7 +209,7 @@ bool ConvertDMRectFromJs(napi_env env, napi_value jsObject, DMRect& rect)
             WLOGFE("[NAPI]Failed to convert parameter to height");
             return false;
         }
-        rect.height_ = height;
+        rect.height_ = static_cast<uint32_t>(height);
     }
     return true;
 }
